refactor: use entities services instead of repos in use cases
Domain ports are XXXService instead of XXXRepository, which allows services to implement more complex methods such as saveEntityAndOtherEntity, leaving technical persistence to repositories defined in infra.
It also allows to remove explicit transaction handling in core use cases, which felt a bit weired as it should not be known to domain whether transactions are or not used.